Moved authentication to be an internal pkg (#1855)
