+++ bindings/azure/servicebusqueues/servicebusqueues.go
@@ -25,8 +25,8 @@ import (
 	"github.com/Azure/go-amqp"
 	backoff "github.com/cenkalti/backoff/v4"
 
-	azauth "github.com/dapr/components-contrib/authentication/azure"
 	"github.com/dapr/components-contrib/bindings"
+	azauth "github.com/dapr/components-contrib/internal/authentication/azure"
 	impl "github.com/dapr/components-contrib/internal/component/azure/servicebus"
 	contrib_metadata "github.com/dapr/components-contrib/metadata"
 	"github.com/dapr/kit/logger"
