+++ internal/authentication/azure/auth.go
