+++ internal/authentication/azure/auth_amqp.go
