+++ internal/authentication/azure/auth_test.go
