+++ internal/authentication/azure/metadata-properties.go
