+++ internal/authentication/azure/storage.go
